+++ components/treemap/src/treemap/index.ts
@@ -1,2 +1,3 @@
+export { ColorMappingsDirective, ColorMappingDirective, ColorMappingsPlugin, ColorMappingPlugin } from './colormapping.directive';
 export { LevelsDirective, LevelDirective, LevelsPlugin, LevelPlugin } from './levels.directive';
 export { TreeMapComponent, TreeMapPlugin } from './treemap.component';