+++ components/treegrid/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## [Unreleased]
 
+## 17.2.39 (2019-07-30)
+
+### TreeGrid
+
+#### Bug Fixes
+
+- #145979 - `checkboxChange` event will trigger for header cell check.
+- #240702 - `allowEditOnDblClick` works fine in Cell Edit mode.
+
+- #145768 - Template element click event is firing when the template column marked as treeColumnIndex
+
+
+#### Breaking Changes
+
+- `checkboxChange` event will trigger after the cell checked.
+
 ## 17.2.35 (2019-07-17)
 
 ### TreeGrid
@@ -59,3 +75,5 @@
 
 - Expanding and Collapsing records is working fine when `pageSizeMode` is set as `All`.
 - `expandAtLevel`, `collapseAtLevel`, `expandAll` and `collapseAll` methods are working fine when `pageSizeMode` is set as `All`.
+
+